Only check for committer ECA license for enabled projects
It can be configured per project whether the ECA validation is enabled
or not. To enable the ECA validation for a project the project must have
the following entry in its `project.config` file in the
`refs/meta/config` branch:

```ini
  [plugin "eca-validation"]
    enabled = true
```

If `plugin.eca-validation.enabled` is not specified in the
`project.config` file, the value is inherited from the parent project.
If it is not set on any parent project the ECA validation is disabled
for this project.

Change-Id: I80795d53dd089524c0ba210e96c303e9dc9efd6f