Use current user as delegate for Gerrit Server identity
The Gerrit Server identity is not an a person that signed
the Eclipse ECA, however, it is the committer of the UI actions
such as the revert.

The server is never acting on its own but is a "delegate" of
the user performing the operation on the UI, hence use the
current identified user as the surrogate identity for the purpose
of the ECA check.

This fix enables the use of the "Revert" action on the Gerrit
UI which otherwise would be impossible to perform, due to the
ECA constraints.

TODO: The Eclipse FDN should validate that this is acceptable
and verify what was the policy adopted on the legacy Gerrit
instance https://git.eclipse.org/r.

Change-Id: I84dac0ae82e08b245aa42f467eebb191e9bd9027