Use the 'github' provider for validating changes against the ECA API
The upgrade of the Eclipse ECA API to Quarkus 3.20 + SDK 1.3.0 has
also, silently removed the 'gerrit' provider in the ValidationRequest.
Therefore, all incoming validations failed with a 400 as raised as
Eclipse Infrastructure Issue 14.

All the GerritHub.io projects are replicated to GitHub anyway, therefore
using the 'github' provider is perfectly valid and also supports the
same parameters.

Bug: https://gitlab.eclipse.org/eclipsefdn/it/websites/api.eclipse.org/-/issues/14
Change-Id: I7734d906dc4d94f7e82e4e653357aa075c0cae3f